% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_srtm.R
\name{get_srtm}
\alias{get_srtm}
\title{Get the SRTM1 elevation data}
\usage{
get_srtm(latitude, longitude)
}
\arguments{
\item{latitude}{a vector of latitude values}

\item{longitude}{a vector of longitude values}
}
\value{
a numeric vector of elevations in meters
}
\description{
Get the SRTM1 elevation data
}
\keyword{internal}
